auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
alembic
python-multipart
pydantic
bcrypt==3.2.0
pydantic-settings